Restore old Play.value behavior
Old fix was wrong and made it impossible to play the 3 of clubs as the value of this play was 0.

Correct fix was to use Card.value instead of Play.value for sorting since Card.value is guaranteed to return a number whereas it makes sense for Play.value to return null for invalid plays.